% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baseballTable}
\alias{baseballTable}
\title{Names and abbreviations of the United States}
\format{A tibble with 27 rows and 5 columns:
\describe{
\item{team}{Team name}
\item{roster}{url of team roster}
\item{stats}{url for cumulative stats}
\item{site_type}{site provider}
}}
\usage{
baseballTable
}
\description{
A data set containing common abbreviations for the 50 United States and
Washington, D.C. Created using \code{Ecfun::\link[Ecfun]{readUSstateAbbreviations}}
}
\keyword{datasets}
