% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterFunctions.R
\name{getData}
\alias{getData}
\title{Title}
\usage{
getData(teamName, rosterURL, statURL, source, sport, output.dir = NULL,
  exportRoster = F, ...)
}
\arguments{
\item{teamName}{}

\item{rosterURL}{}

\item{source}{}

\item{sport}{}

\item{statsURL}{}
}
\description{
Title
}
