% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterFunctions.R
\name{getRoster}
\alias{getRoster}
\title{Title}
\usage{
getRoster(
  teamName,
  url,
  source = c("sidearm", "wmt", "neulion", "presto", "liberty"),
  sport = c("baseball")
)
}
\arguments{
\item{sport}{}
}
\value{

}
\description{
Title
}
