% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{getSchedule}
\alias{getSchedule}
\title{Get schedule for a single sport}
\usage{
getSchedule(team, year)
}
\arguments{
\item{team}{name of sport}

\item{year}{numeric, four-digit year}
}
\value{
data frame with four columns for the opponent, game location, game date, and opponent url
}
\description{
Get schedule for a single sport
}
