% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterFunctions.R
\name{getStats}
\alias{getStats}
\title{Title}
\usage{
getStats(team, statURL, roster_df, col.names = format, ...)
}
\arguments{
\item{team}{}

\item{statURL}{}

\item{roster_df}{}

\item{...}{}
}
\value{

}
\description{
Title
}
