% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule_details}
\alias{schedule_details}
\title{Parse schedule details}
\usage{
schedule_details(game_node)
}
\arguments{
\item{sport}{name of sport}

\item{year}{numeric, four-digit year}
}
\value{
vector with team name, game location, game date, and opponent url
}
\description{
Parse schedule details
}
