% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{Names and abbreviations of the United States}
\format{A tibble with 51 rows and 4 variables:
\describe{
\item{Name}{State name}
\item{USPS}{Two-letter postal abbreviation}
\item{AP}{Abbreviation as used by the Associated Press}
\item{Other}{Other, non-standard abbreviations}
...
}}
\source{
\url{https://en.wikipedia.org/wiki/List_of_U.S._state_abbreviations}
}
\usage{
states
}
\description{
A data set containing common abbreviations for the 50 United States and
Washington, D.C. Created using \code{Ecfun::\link[Ecfun]{readUSstateAbbreviations}}
}
\keyword{datasets}
