% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterFunctions.R
\name{getRoster}
\alias{getRoster}
\title{Title}
\usage{
getRoster(teamName, url, source, sport, ...)
}
\arguments{
\item{...}{Additional Arguments passed to Sidearm Roster Functions}
}
\description{
Title
}
