% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{Names and abbreviations of the United States}
\format{
A tibble with 146 rows and 2 variables:
\describe{
\item{USPS}{Two-letter postal abbreviation}
\item{state}{Other abbreviations and full name of the state}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_U.S._state_abbreviations}
}
\usage{
states
}
\description{
A data set containing common abbreviations for the 50 United States and
Washington, D.C. Created using \code{Ecfun::\link[Ecfun]{readUSstateAbbreviations}}
}
\keyword{datasets}
